+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.4] - 2025-05-23
+
 - hnefatafl-client: don't limit the valid password characters.
 - fix the lowercasing of password bug.
 
@@ -268,7 +270,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - hnefatafl-client: Make users and games scrollable.
 - Make all the features dependencies.
 
-[unreleased]: https://github.com/dcampbell24/hnefatafl-copenhagen/compare/v1.1.3...main
+[unreleased]: https://github.com/dcampbell24/hnefatafl-copenhagen/compare/v1.1.4...main
+[1.1.4]:  https://github.com/dcampbell24/hnefatafl-copenhagen/compare/v01.1.3...v1.1.4
 [1.1.3]:  https://github.com/dcampbell24/hnefatafl-copenhagen/compare/v01.1.2...v1.1.3
 [1.1.2]:  https://github.com/dcampbell24/hnefatafl-copenhagen/compare/v01.1.1...v1.1.2
 [1.1.1]:  https://github.com/dcampbell24/hnefatafl-copenhagen/compare/v01.1.0...v1.1.1